migrating to cmake

+++ Emulator/Agnus/Copper/Copper.cpp
@@ -12,7 +12,7 @@
 #include "Amiga.h"
 #include "CopperDebugger.h"
 #include "Checksum.h"
-#include "IO.h"
+#include "IOUtils.h"
 #include "PixelEngine.h"
 
 Copper::Copper(Amiga& ref) : SubComponent(ref)
@@ -126,7 +126,7 @@ Copper::findMatch(Beam &match) const
 
     // Iterate through all lines starting from the current position
     isize numLines = agnus.frame.numLines();
-    while ((beam >> 8) < numLines) {
+    while ((isize)(beam >> 8) < numLines) {
 
         // Check if the vertical components are equal
         if ((beam & mask & ~0xFF) == (comp & mask & ~0xFF)) {
